Resets between games / Popmessage for wave
If you play multiple games then the plugin resets the globals
so that stats work properly in following game.

Added popmessage for Asteroids wave, so that user knows the next
incoming wave. There's a small bit of logic to NOT display this if
the number of ships is down to 0.

Finally, the info output has been updated.