Fixes bug by moving "ship score" block
This fixes a bug in which a ship after spawning is immediately
hit (in this case with a small rock). Before this commit, the
100 points would go to 'the next ship'. This change makes it so
the current ship gets the points.

    Number of ships changed: BEFORE: 3 AFTER: 2
    Actual Score: 16390 Diff: 100
    Number of ships changed: BEFORE: 2 AFTER: 1

Also, per the output above, I changed the output so that the
player knows what's going with the "ships changed". The count
can increase (in the case of reaching a 10000 point increment) or
the count can decrease (in the case of a ship getting killed).