Prints Statistics Table at End of Game (STDOUT) (#2)
* Adds print_score_board() function

This prints the in-game table to the STDOUT for any post-game uses.
This was coded with some help from AI (which suggested where to place
this function so that it was visible to the process_frame() function).
I also followed AI guidance and moved the draw_score_board 
function "to the top" of the file.

* Brings in 'Unified Renderer'

This is some code consolidation, thanks to AI.

* Updates README